+++ commons-cache/src/main/java/com/hb0730/commons/cache/support/serial/Serializer.java
@@ -1,5 +1,7 @@
 package com.hb0730.commons.cache.support.serial;
 
+import com.hb0730.commons.cache.exception.SerializationException;
+
 import javax.annotation.Nullable;
 
 /**
@@ -17,10 +19,10 @@ public interface Serializer {
      *
      * @param obj 序列化对象
      * @return 二进制
-     * @throws Exception 序列号异常
+     * @throws SerializationException 序列化异常
      */
     @Nullable
-    byte[] serialize(@Nullable Object obj) throws Exception;
+    byte[] serialize(@Nullable Object obj) throws SerializationException;
 
 
     /**
@@ -28,9 +30,9 @@ public interface Serializer {
      *
      * @param bytes 字节数组
      * @return 序列化对象
-     * @throws Exception 反序列化异常
+     * @throws SerializationException 反序列化异常
      */
     @Nullable
-    Object deserialize(@Nullable byte[] bytes) throws Exception;
+    Object deserialize(@Nullable byte[] bytes) throws SerializationException;
 
 }